+++ modules/AdminToolbox/ChangeLog.md
@@ -182,6 +182,7 @@
 * **10.2.7** Update required modules
 * **10.2.8** Increment Manifest to resolve missing manifest version in github timeline
 * **10.2.9** Update required modules
+* **10.2.10** Update required modules
 
 #The line below this one is used strictly by a github action. That action is responsible for publishing new versions of this module to Powershell Gallery, upon pushing an iterative version to Github.
-10.2.9
+10.2.10